+++ Ejercicio 1_Miercoles_21-Sep/Exe02/exercise02.cpp
@@ -1,0 +1,78 @@
+// =================================================================
+//
+// File: exercise02.cpp
+// Author(s): Diego Alfonso Ramirez Montes
+// Description: This file contains the code that performs the sum of
+//				all prime numbers less than or equal to MAXIMUM. The 
+//				time this implementation takes will be used as the 
+//				basis to calculate the improvement obtained with 
+//				parallel technologies.
+//
+// Copyright (c) 2022 by Tecnologico de Monterrey.
+// All Rights Reserved. May be reproduced for any non-commercial
+// purpose.
+//
+// =================================================================
+
+#include <iostream>
+#include <iomanip>
+#include <cstring>
+#include <cmath>
+#include <algorithm>
+#include "utils.h"
+
+#define MAXIMUM 1000000 //1e6
+
+using namespace std;
+
+bool comp(int a){
+	int i;
+	
+	if(a < 2){
+		return false;
+		
+	}
+	for(i = 2; i < sqrt(a); i++){
+		if(a % i == 0){
+			return false;
+		
+		}
+	}
+	return true;
+	
+}
+
+double sumP(int size){
+	double acum;
+	int i;
+	
+	acum = 0;
+	for(i = 2; i < size; i++){
+		if(comp(i)){
+			acum += i;
+			
+		}
+		
+	}
+	return acum;
+	
+}
+
+int main(int argc, char* argv[]) {
+	int i;
+	double ms, result;
+
+	cout << "Starting..." << endl;
+	ms = 0;
+	for (int i = 0; i < N; i++) {
+		start_timer();
+
+		result = sumP(MAXIMUM);
+
+		ms += stop_timer();
+	}
+	cout << "result = " << fixed << setprecision(0) << result << "\n";
+	cout << "avg time = " << setprecision(5) << (ms / N) << " ms" << endl;
+
+	return 0;
+}
