+++ Ejercicio 1_Miercoles_21-Sep/Exe03/exercise03.cpp
@@ -1,0 +1,50 @@
+// =================================================================
+//
+// File: exercise03.cpp
+// Author(s): Diego Alfonso Ramirez Montes
+// Description: This file contains the code that implements the
+//				enumeration sort algorithm. The time this implementation
+//				takes ill be used as the basis to calculate the
+//				improvement obtained with parallel technologies.
+//
+// Copyright (c) 2022 by Tecnologico de Monterrey.
+// All Rights Reserved. May be reproduced for any non-commercial
+// purpose.
+//
+// =================================================================
+
+#include <iostream>
+#include <iomanip>
+#include <cstring>
+#include "utils.h"
+
+const int SIZE = 10000; //1e4
+
+using namespace std;
+
+
+
+int main(int argc, char* argv[]) {
+	int *a;
+	double ms;
+
+	a = new int[SIZE];
+	random_array(a, SIZE);
+	display_array("before", a);
+
+	cout << "Starting..." << endl;
+	ms = 0;
+	// create object here
+	for (int i = 0; i < N; i++) {
+		start_timer();
+
+		// call your code here.
+
+		ms += stop_timer();
+	}
+	display_array("after", b);
+	cout << "avg time = " << setprecision(5) << (ms / N) << " ms" << endl;
+
+	delete [] a;
+	return 0;
+}
